一時動かなくなっていたフィルター機能を 別ファイルに分離して復活
